requirements.txt
# Core Dependencies
streamlit>=1.28.0
openai>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0
Pillow>=10.0.0

# Optional Development Tools
black>=23.0.0
pytest>=7.4.0
python-dotenv>=1.0.0